refactor/#277: enum -> constant로 변경
