+++ constants/Auth.ts
@@ -1,0 +1,5 @@
+export const LoginUserState = {
+  NO_USER: 'noUser',
+  WRONG_PLATFORM: 'wrongPlatform',
+  INVITE_MISMATCH: 'inviteMismatch',
+} as const;
