+++ types/auth.ts
@@ -20,9 +20,3 @@ export interface jwtAccessTokenResponse {
   grantType: string;
   refreshToken: string;
 }
-
-export enum LoginUserState {
-  NO_USER = 'noUser',
-  WRONG_PLATFORM = 'wrongPlatform',
-  INVITE_MISMATCH = 'inviteMismatch',
-}
